newslate checks for windows 8 and deletes properly

+++ toolCloudProject/newSlate.py
@@ -14,6 +14,8 @@ import sys
 import platform
 from subprocess import *
 
+
+
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "toolCloudProject.settings")
 
 from django.conf import settings
@@ -50,7 +52,10 @@ print ("-->	Static files collected")
 
 
 print ("-->	Deleting old database file...")
-os.system('rm db.sqlite3')
+if platform.release() == 8:					#checks to see what windows version is running
+	os.system('del db.sqlite3')
+else:
+	os.system('rm db.sqlite3')
 print ("-->	Delete Complete")
 
 print ("")
